Rename define-sealable-generic-function to defgeneric.
We intentionally shadow a standard CL macro.  This is not just to teach
users of this library the lesson to not just :use other libraries, but to
indicate that the new sealable-metaobjects:defgeneric macro has exactly the
same semantics as cl:defgeneric, except that it permits non-exhaustive
arguments to the :argument-precedence-order option, and that it attempts to
set the specializer profile of the generic function to a meaningful value.